% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_org.R
\name{dataOrganization}
\alias{dataOrganization}
\title{Title}
\usage{
dataOrganization(data, target, testDataLength = nrow(data))
}
\arguments{
\item{data}{data to be organized}

\item{target}{"Appliances" or "lights"}

\item{testDataLength}{length of output subset (from 1 do testDataLength)}
}
\value{
organised data set (or subset, if testDataLength specified),
}
\description{
Title
}
