% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_eval.R
\name{model_eval}
\alias{model_eval}
\title{Model Evaluation}
\usage{
model_eval(test_data, fun, attr, crossval_number)
}
\arguments{
\item{test_data}{}

\item{fun}{}

\item{attr}{}

\item{crossval_number}{}
}
\value{
data frame with measures in following order: MSE, RRSE, MAE, RMSE, RAE, COR, RSE
}
\description{
Model Evaluation
}
