% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.filter.R
\name{simple.filter}
\alias{simple.filter}
\title{Simple attribute filter}
\usage{
simple.filter(formula, data, dd = dd.chi2, cd = cd.kruskal, cc = cc.spearman)
}
\arguments{
\item{formula}{}

\item{data}{}

\item{dd}{}

\item{cd}{}

\item{cc}{}
}
\value{

}
\description{
Simple attribute filter
}
