% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{feature_selection}
\alias{feature_selection}
\title{Feature Selection}
\usage{
feature_selection(test_data, type, part)
}
\arguments{
\item{test_data}{}

\item{type}{}

\item{part}{}
}
\value{
formula with selected attributes
}
\description{
Feature Selection
}
