% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{feature_selection}
\alias{feature_selection}
\title{Feature Selection}
\usage{
feature_selection(test_data, type = "rf", part = 1, trees_num = 10)
}
\arguments{
\item{test_data}{- data to perform feature ranking}

\item{type}{- "rf" - randomForest IMPORTANCE-based ranking (MSE)(default),
"bootstrap" - bootstraped R2-based (coef. of determination) ranking
"simple" - simple filter based algorithms}

\item{part}{- part of attributes returned by feature selection, 0 < part <=1 (default=0.5)}

\item{trees_num}{- numbers of trees (randomForest) or bootstrap sets (bootstrap)}
}
\value{
formula with selected attributes
}
\description{
Feature Selection
}
