% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.filter.R
\name{simple.filter}
\alias{simple.filter}
\title{Simple attribute filter}
\usage{
simple.filter(formula, data, dd = dd.chi2, cd = cd.kruskal, cc = cc.spearman)
}
\arguments{
\item{formula}{}

\item{data}{data to perform feature selection}

\item{dd}{test to perform selection with discrete attributes and discrete target}

\item{cd}{test to perform selection with continuous attributes and discrete target}

\item{cc}{test to perform selection with continuous attributes and discrete target}
}
\value{
named by attribute list of test measures
}
\description{
Simple attribute filter
}
